Diagram._add_morphism_closure: Remove extra flags.
Previously, when Diagram would contain both premises and conclusions, it
would use slightly different logic for computing the closure of premises
and conclusions.  The behaviour was controlled by the two arguments:
recurse_composites and add_identities.  Now that Diagram only stores a
set of morphisms, it is not necessary to have this difference.