Embeddings: Transition to simple graphs.
Diagrams are essentially multigraphs.  To be able to list all embeddings,
we will need to first convert both the pattern and the model to a simple
graph, then apply Ullman's algorithm (see references in the code) to
generate all subgraph embeddings, and then, from these embeddings, generate
the multigraph embeddings.

To apply Ullman's algorithm, we will need to do matrix multiplication
with adjacency matrices, this is why we set up bidirectional mappings
between the objects of the diagrams and their indices.

This commit introduces the main magic which will eventually allow
``diagram_embeddings`` to work with infinite diagrams.