Embeddings: Improve the morphism property matching strategy.
Previously, ``match_morphisms`` would only consider two morphisms to
match if they had exactly the same set of properties.  This commit
changes this behaviour; a morphism from the pattern now matches a
morphism from the model if its properties in the pattern are a subset of
the properties in the model.

Apparently, this is more consistent with the idea of pattern matching in
general.  However, this change is more of an intuitive one.  If it
doesn't work out well, I will revert it.