Diagram: Implement Kosaraju's algorithm to detect loops.
Previously, ``Diagram`` used an adapted version of topological sort to
detect loops.  That solution would stop at detecting any loop.  Yet, it
is sometimes necessary to know whether a certain hom-set of a diagram
is finite or not.  To answer this question, I have implemented
Kosaraju's algorithm for finding the strongly connected components of
the graph defined by the generators.

This commit does not add the possibility to check whether a certain
hom-set is finite, though.  Instead, it only introduces Kosaraju's
algorithm.