Diagram: Forbid composites whose components are not generators.
Previously ``Diagram`` would tacitly accept such composites.  However,
since all composites are always represented in a flattened down form,
this would result in tricky situations.

The limitation condition introduced by this commit is not in fact a
serious restriction.  Any composite morphism can be flattened down to a
named morphism or, alternatively, one can just explicitly add the
components of the composite to the diagram.