Embeddings: Generate actual diagram embeddings.
This commit finalises the functionality of ``diagram_embeddings`` by
making this function provide complete embeddings.  Note again that for
an infinite pattern only the information about the embeddings of
expanded generators of the pattern will be provided.  (See the dosctring
of ``Diagram`` for more details.)

This commit makes the tests pass.