Start implementing and write a basic test for commutativity inference.
This commit adds the function ``_check_commutativity_with_diagrams`` and
a trivial test for it.  This function will eventually be able to carry out
the commutativity stage of the inference process, i.e., see whether a
diagram is commutative basing on a list of diagrams known to be
commutative.  In that context, more extensive tests will be written.